removed gsPPN.cpp
